Moved file location of crash log dumps to <app_files_root>/logs/<log_name>. Updated ITimespanDbService.

+++ Timespan.Database/Services/Interfaces/ITimespanDbService.cs
@@ -12,7 +12,7 @@ public interface ITimespanDbService {
 
 	public Task<List<Types.Task>> QueryTasksAsync();
 
-	public Task<Types.Task?> QueryTasksByIdAsync(int id);
+	public Task<Types.Task?> QueryTasksByIdAsync(long id);
 
 	public Task<Types.Task?> QueryCurrentTaskAsync();
